[fix] 회원가입 및 로그인 방식 수정
