--- backend/landing/urls.py
@@ -1,6 +1,0 @@
-from django.urls import path
-from .views import ServiceInfoView
-
-urlpatterns = [
-    path('service-info/', ServiceInfoView.as_view(), name='service_info'),
-]
