--- backend/user/serializers.py
@@ -1,15 +1,0 @@
-from rest_framework import serializers
-from .models import User
-from django.contrib.auth.hashers import make_password
-
-class UserSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = User
-        fields = ['id', 'username', 'personal_id', 'password']
-        extra_kwargs = {
-            'password': {'write_only': True}
-        }
-
-    def create(self, validated_data):
-        validated_data['password'] = make_password(validated_data['password'])  # 비밀번호 암호화
-        return super().create(validated_data)
